backend/requirements.txt
acres==0.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
ci-info==0.4.0
click==8.3.1
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
cors==1.0.1
cryptography==46.0.5
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
etelemetry==0.3.1
fastapi==0.135.1
filelock==3.25.1
future==1.0.0
gevent==25.9.1
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
jwt==1.4.0
logger==1.4
looseversion==1.3.0
lxml==6.0.2
motor==3.7.1
networkx==3.6.1
nibabel==5.4.2
nipype==1.11.0
numpy==2.4.3
packaging==26.0
pandas==3.0.1
passlib==1.7.4
pathlib==1.0.1
prov==2.1.1
puremagic==2.1.0
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydot==4.0.1
pymongo==4.16.0
PyMuPDF==1.27.2.2
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
pyxnat==1.6.4
rdflib==7.6.0
requests==2.32.5
requests-file==3.0.1
rsa==4.9.1
scipy==1.17.1
simplejson==3.20.2
six==1.17.0
soupsieve==2.8.3
starlette==0.52.1
tldextract==5.3.1
traits==7.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
zope.event==6.1
zope.interface==8.2
openai
qdrant-client
